# node/attendanceMgm/api/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "TML app",
    "description": "APIs for TML app",
    "version": "1.0"
  },
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json",
    "text/xml"
  ],
  "host": "localhost:8080",
  "basePath": "/v1/tml",
  "paths": {    
...
      "/alerts":{  
        "post": {
        "x-swagger-router-controller": "AlertController",
        "operationId": "sendAlert",
        "tags": [
          "/alerts"
        ],
        "description": "sendAlert",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "alertObject",
            "description": "sendAlert",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Alert"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Alert"
            }
          },
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
	  "get":{  
            "x-swagger-router-controller":"AlertController",
            "operationId":"getAlerts",
            "tags":[  
               "/alerts"
            ],
            "description":"Returns all alerts",
            "parameters":[ 
			    {
            "name": "status",
            "in": "query",
            "description": "status",
            "required": false,
            "type": "string"
          }                
            ],
            "responses":{  
               "200":{  
                  "description":"Successful request.",
                  "schema":{  
                     "$ref":"#/definitions/Alert"
                  }
               },
               "default":{  
                  "description":"Invalid request.",
                  "schema":{  
                     "$ref":"#/definitions/Error"
                  }
               }
            }
         },
		 "options": {
        "x-swagger-router-controller": "AlertController",
        "operationId": "handleOption",
        "tags": [
          "/alerts"
        ],
        "description": "send alert",		
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
		 
      },
   "/alerts/{id}": {
      "put": {
        "x-swagger-router-controller": "AlertController",
        "operationId": "putAlert",
        "tags": [
          "/alerts"
        ],
        "description": "update alert",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "alert id",
            "required": true,
            "in": "path",
            "type":"string"
          },
          {
            "name": "alertObject",
            "description": "update Entity",
            "required": true,
            "in": "body",
            "schema": {
              "$ref": "#/definitions/Alert_PUT"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful operation",
            "schema": {
              "$ref": "#/definitions/Alert"
            }
          },
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      },
      "options": {
        "x-swagger-router-controller": "AlertController",
        "operationId": "handleOption",
        "tags": [
          "/alerts"
        ],
        "parameters": [
          {
            "name": "id",
            "description": "alert id",
            "required": true,
            "in": "path",
            "type":"string"
          }
        ],
        "description": "update alert",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "200": {
            "description": "Successful operation"
          },
          "default": {
            "description": "Invalid request.",
            "schema": {
              "$ref": "#/definitions/Error"
            }
          }
        }
      }
    }	  
  },
  "definitions": {
    "Error": {
      "properties": {
        "message": {
          "type": "string"
        }
      },
      "required": [
        "message"
      ]
    },    
...
    "Alert": {
      "properties": {
        "alerts_text": {
          "type": "string"
        },
        "alert_zone": {
          "type": "string"
        },
		"created_by":{
		"type":"string"
		}
      }
    },	
...
	"Alert_PUT": {
      "properties": {
        "status": {
          "type": "string"
        }        
      }
    }
  }
}